Move recorder statistics API to data/recorder.ts
