+++ src/data/history.ts
@@ -2,7 +2,6 @@ import { HassEntities, HassEntity } from "home-assistant-js-websocket";
 import { computeDomain } from "../common/entity/compute_domain";
 import { computeStateDisplayFromEntityAttributes } from "../common/entity/compute_state_display";
 import {
-  computeStateName,
   computeStateNameFromEntityAttributes,
 } from "../common/entity/compute_state_name";
 import { LocalizeFunc } from "../common/translations/localize";
@@ -61,87 +60,6 @@ export interface TimelineEntity {
 export interface HistoryResult {
   line: LineChartUnit[];
   timeline: TimelineEntity[];
-}
-
-export type StatisticType = "sum" | "min" | "max" | "mean";
-
-export interface Statistics {
-  [statisticId: string]: StatisticValue[];
-}
-
-export interface StatisticValue {
-  statistic_id: string;
-  start: string;
-  end: string;
-  last_reset: string | null;
-  max: number | null;
-  mean: number | null;
-  min: number | null;
-  sum: number | null;
-  state: number | null;
-}
-
-export interface StatisticsMetaData {
-  display_unit_of_measurement: string;
-  statistics_unit_of_measurement: string;
-  statistic_id: string;
-  source: string;
-  name?: string | null;
-  has_sum: boolean;
-  has_mean: boolean;
-}
-
-export type StatisticsValidationResult =
-  | StatisticsValidationResultNoState
-  | StatisticsValidationResultEntityNotRecorded
-  | StatisticsValidationResultEntityNoLongerRecorded
-  | StatisticsValidationResultUnsupportedStateClass
-  | StatisticsValidationResultUnitsChanged
-  | StatisticsValidationResultUnsupportedUnitMetadata
-  | StatisticsValidationResultUnsupportedUnitState;
-
-export interface StatisticsValidationResultNoState {
-  type: "no_state";
-  data: { statistic_id: string };
-}
-
-export interface StatisticsValidationResultEntityNoLongerRecorded {
-  type: "entity_no_longer_recorded";
-  data: { statistic_id: string };
-}
-
-export interface StatisticsValidationResultEntityNotRecorded {
-  type: "entity_not_recorded";
-  data: { statistic_id: string };
-}
-
-export interface StatisticsValidationResultUnsupportedStateClass {
-  type: "unsupported_state_class";
-  data: { statistic_id: string; state_class: string };
-}
-
-export interface StatisticsValidationResultUnitsChanged {
-  type: "units_changed";
-  data: { statistic_id: string; state_unit: string; metadata_unit: string };
-}
-
-export interface StatisticsValidationResultUnsupportedUnitMetadata {
-  type: "unsupported_unit_metadata";
-  data: {
-    statistic_id: string;
-    device_class: string;
-    metadata_unit: string;
-    supported_unit: string;
-  };
-}
-
-export interface StatisticsValidationResultUnsupportedUnitState {
-  type: "unsupported_unit_state";
-  data: { statistic_id: string; device_class: string; metadata_unit: string };
-}
-
-export interface StatisticsValidationResults {
-  [statisticId: string]: StatisticsValidationResult[];
 }
 
 export interface HistoryStates {
@@ -448,133 +366,4 @@ export const computeHistory = (
   );
 
   return { line: unitStates, timeline: timelineDevices };
-};
-
-// Statistics
-
-export const getStatisticIds = (
-  hass: HomeAssistant,
-  statistic_type?: "mean" | "sum"
-) =>
-  hass.callWS<StatisticsMetaData[]>({
-    type: "recorder/list_statistic_ids",
-    statistic_type,
-  });
-
-export const getStatisticMetadata = (
-  hass: HomeAssistant,
-  statistic_ids?: string[]
-) =>
-  hass.callWS<StatisticsMetaData[]>({
-    type: "recorder/get_statistics_metadata",
-    statistic_ids,
-  });
-
-export const fetchStatistics = (
-  hass: HomeAssistant,
-  startTime: Date,
-  endTime?: Date,
-  statistic_ids?: string[],
-  period: "5minute" | "hour" | "day" | "month" = "hour"
-) =>
-  hass.callWS<Statistics>({
-    type: "recorder/statistics_during_period",
-    start_time: startTime.toISOString(),
-    end_time: endTime?.toISOString(),
-    statistic_ids,
-    period,
-  });
-
-export const validateStatistics = (hass: HomeAssistant) =>
-  hass.callWS<StatisticsValidationResults>({
-    type: "recorder/validate_statistics",
-  });
-
-export const updateStatisticsMetadata = (
-  hass: HomeAssistant,
-  statistic_id: string,
-  unit_of_measurement: string | null
-) =>
-  hass.callWS<void>({
-    type: "recorder/update_statistics_metadata",
-    statistic_id,
-    unit_of_measurement,
-  });
-
-export const clearStatistics = (hass: HomeAssistant, statistic_ids: string[]) =>
-  hass.callWS<void>({
-    type: "recorder/clear_statistics",
-    statistic_ids,
-  });
-
-export const calculateStatisticSumGrowth = (
-  values: StatisticValue[]
-): number | null => {
-  if (!values || values.length < 2) {
-    return null;
-  }
-  const endSum = values[values.length - 1].sum;
-  if (endSum === null) {
-    return null;
-  }
-  const startSum = values[0].sum;
-  if (startSum === null) {
-    return endSum;
-  }
-  return endSum - startSum;
-};
-
-export const calculateStatisticsSumGrowth = (
-  data: Statistics,
-  stats: string[]
-): number | null => {
-  let totalGrowth: number | null = null;
-
-  for (const stat of stats) {
-    if (!(stat in data)) {
-      continue;
-    }
-    const statGrowth = calculateStatisticSumGrowth(data[stat]);
-
-    if (statGrowth === null) {
-      continue;
-    }
-    if (totalGrowth === null) {
-      totalGrowth = statGrowth;
-    } else {
-      totalGrowth += statGrowth;
-    }
-  }
-
-  return totalGrowth;
-};
-
-export const statisticsHaveType = (
-  stats: StatisticValue[],
-  type: StatisticType
-) => stats.some((stat) => stat[type] !== null);
-
-export const adjustStatisticsSum = (
-  hass: HomeAssistant,
-  statistic_id: string,
-  start_time: string,
-  adjustment: number
-): Promise<void> =>
-  hass.callWS({
-    type: "recorder/adjust_sum_statistics",
-    statistic_id,
-    start_time,
-    adjustment,
-  });
-
-export const getStatisticLabel = (
-  hass: HomeAssistant,
-  statisticsId: string,
-  statisticsMetaData: StatisticsMetaData | undefined
-): string => {
-  const entity = hass.states[statisticsId];
-  if (entity) {
-    return computeStateName(entity);
-  }
-  return statisticsMetaData?.name || statisticsId;
 };
