+++ src/panels/config/energy/components/ha-energy-device-settings.ts
@@ -15,7 +15,7 @@ import {
 import {
   StatisticsMetaData,
   getStatisticLabel,
-} from "../../../../data/history";
+} from "../../../../data/recorder";
 import {
   showAlertDialog,
   showConfirmationDialog,
