+++ src/panels/config/energy/components/ha-energy-gas-settings.ts
@@ -16,7 +16,7 @@ import {
 import {
   StatisticsMetaData,
   getStatisticLabel,
-} from "../../../../data/history";
+} from "../../../../data/recorder";
 import {
   showAlertDialog,
   showConfirmationDialog,
