+++ src/panels/developer-tools/statistics/show-dialog-statistics-adjust-sum.ts
@@ -1,5 +1,5 @@
 import { fireEvent } from "../../../common/dom/fire_event";
-import { StatisticsMetaData } from "../../../data/history";
+import { StatisticsMetaData } from "../../../data/recorder";
 
 export const loadAdjustSumDialog = () =>
   import("./dialog-statistics-adjust-sum");
