+++ src/panels/lovelace/cards/energy/hui-energy-grid-neutrality-gauge-card.ts
@@ -13,7 +13,7 @@ import {
   getEnergyDataCollection,
   GridSourceTypeEnergyPreference,
 } from "../../../../data/energy";
-import { calculateStatisticsSumGrowth } from "../../../../data/history";
+import { calculateStatisticsSumGrowth } from "../../../../data/recorder";
 import { SubscribeMixin } from "../../../../mixins/subscribe-mixin";
 import type { HomeAssistant } from "../../../../types";
 import type { LovelaceCard } from "../../types";
