+++ src/panels/lovelace/cards/energy/hui-energy-solar-graph-card.ts
@@ -43,7 +43,7 @@ import {
   Statistics,
   StatisticsMetaData,
   getStatisticLabel,
-} from "../../../../data/history";
+} from "../../../../data/recorder";
 import { FrontendLocaleData } from "../../../../data/translation";
 import { SubscribeMixin } from "../../../../mixins/subscribe-mixin";
 import { HomeAssistant } from "../../../../types";
