+++ src/panels/lovelace/cards/types.ts
@@ -1,4 +1,4 @@
-import { StatisticType } from "../../../data/history";
+import { StatisticType } from "../../../data/recorder";
 import { ActionConfig, LovelaceCardConfig } from "../../../data/lovelace";
 import { FullCalendarView, TranslationDict } from "../../../types";
 import { Condition } from "../common/validate-condition";
